Move function raster2split to module snail_intersections
